Remove `BaseKeyChain` (#9830)
* Remove BaseKeyChain
* Extract static methods
