+++ WalletWasabi/WabiSabi/Client/KeyChain.cs
@@ -1,31 +1,35 @@
-using System.Collections.Generic;
 using NBitcoin;
+using System.Collections.Generic;
 using System.Linq;
 using WalletWasabi.Blockchain.Keys;
+using WalletWasabi.Crypto;
+using WalletWasabi.Extensions;
 using WalletWasabi.Wallets;
-using System.Diagnostics.CodeAnalysis;
 
 namespace WalletWasabi.WabiSabi.Client;
 
-public class KeyChain : BaseKeyChain
+public class KeyChain : IKeyChain
 {
-	public KeyChain(KeyManager keyManager, Kitchen kitchen) : base(kitchen)
+	public KeyChain(KeyManager keyManager, Kitchen kitchen)
 	{
 		if (keyManager.IsWatchOnly)
 		{
-			throw new ArgumentException("A watch-only keymanager cannot be used to initialize a keychain.");
+			throw new ArgumentException("A watch-only key manager cannot be used to initialize a key chain.");
 		}
+
 		KeyManager = keyManager;
+		Kitchen = kitchen;
 	}
 
 	private KeyManager KeyManager { get; }
+	private Kitchen Kitchen { get; }
 
-	protected override Key GetMasterKey()
+	private Key GetMasterKey()
 	{
 		return KeyManager.GetMasterExtKey(Kitchen.SaltSoup()).PrivateKey;
 	}
 
-	public override void TrySetScriptStates(KeyState state, IEnumerable<Script> scripts)
+	public void TrySetScriptStates(KeyState state, IEnumerable<Script> scripts)
 	{
 		foreach (var hdPubKey in KeyManager.GetKeys(key => scripts.Any(key.ContainsScript)))
 		{
@@ -33,26 +37,52 @@ public class KeyChain : BaseKeyChain
 		}
 	}
 
-	protected override BitcoinSecret GetBitcoinSecret(Script scriptPubKey)
+	public OwnershipProof GetOwnershipProof(IDestination destination, CoinJoinInputCommitmentData commitmentData)
 	{
-		var hdKey = KeyManager.GetSecrets(Kitchen.SaltSoup(), scriptPubKey).Single();
+		ExtKey? hdKey = KeyManager.GetSecrets(Kitchen.SaltSoup(), destination.ScriptPubKey).SingleOrDefault();
+
 		if (hdKey is null)
 		{
-			throw new InvalidOperationException($"The signing key for '{scriptPubKey}' was not found.");
+			throw new InvalidOperationException($"The signing key for '{destination.ScriptPubKey}' was not found.");
 		}
 
-		var derivedScriptPubKeyType = scriptPubKey switch
-		{
-			_ when scriptPubKey.IsScriptType(ScriptType.P2WPKH) => ScriptPubKeyType.Segwit,
-			_ when scriptPubKey.IsScriptType(ScriptType.Taproot) => ScriptPubKeyType.TaprootBIP86,
-			_ => throw new NotSupportedException("Not supported script type.")
-		};
+		Key masterKey = GetMasterKey();
+		BitcoinSecret secret = hdKey.GetBitcoinSecret(KeyManager.GetNetwork(), destination.ScriptPubKey);
 
-		if (hdKey.PrivateKey.PubKey.GetScriptPubKey(derivedScriptPubKeyType) != scriptPubKey)
+		return NBitcoinExtensions.GetOwnershipProof(masterKey, secret, destination.ScriptPubKey, commitmentData);
+	}
+
+	public Transaction Sign(Transaction transaction, Coin coin, PrecomputedTransactionData precomputedTransactionData)
+	{
+		transaction = transaction.Clone();
+
+		if (transaction.Inputs.Count == 0)
 		{
-			throw new InvalidOperationException("The key cannot generate the utxo scriptpubkey. This could happen if the wallet password is not the correct one.");
+			throw new ArgumentException("No inputs to sign.", nameof(transaction));
 		}
-		var secret = hdKey.PrivateKey.GetBitcoinSecret(KeyManager.GetNetwork());
-		return secret;
+
+		var txInput = transaction.Inputs.AsIndexedInputs().FirstOrDefault(input => input.PrevOut == coin.Outpoint);
+
+		if (txInput is null)
+		{
+			throw new InvalidOperationException("Missing input.");
+		}
+
+		ExtKey? hdKey = KeyManager.GetSecrets(Kitchen.SaltSoup(), coin.ScriptPubKey).SingleOrDefault();
+
+		if (hdKey is null)
+		{
+			throw new InvalidOperationException($"The signing key for '{coin.ScriptPubKey}' was not found.");
+		}
+
+		BitcoinSecret secret = hdKey.GetBitcoinSecret(KeyManager.GetNetwork(), coin.ScriptPubKey);
+
+		TransactionBuilder builder = Network.Main.CreateTransactionBuilder();
+		builder.AddKeys(secret);
+		builder.AddCoins(coin);
+		builder.SetSigningOptions(new SigningOptions(TaprootSigHash.All, (TaprootReadyPrecomputedTransactionData)precomputedTransactionData));
+		builder.SignTransactionInPlace(transaction);
+
+		return transaction;
 	}
 }
